+++ src/modules/auth/pages/login/index.js
@@ -3,6 +3,8 @@ import './style.css'
 import '../../../../components/input'
 import '../../../../components/button'
 import Component from '../../../../lib/dom/Component'
+import Button from '../../../../components/ui/button'
+import { loading } from '../../../../lib/helpers/components'
 
 const loginBtnId = 'login-submit'
 const registerBtnId = 'login-new-user'
@@ -17,18 +19,76 @@ class LoginPage extends Component {
     super(options)
   }
 
-  async mounted() { return import('./mounted') }
+  async mounted() {
+    console.log('render login page')
+
+    const loginBtnId = 'login-submit'
+    const registerBtnId = 'login-new-user'
+    const loginLoadingId = 'login-submit-loading'
+
+    const submitBtn = document.getElementById(loginBtnId)
+
+    submitBtn.addEventListener('click', async (e) => {
+      e.preventDefault()
+      const anchor = e.target.closest('a')
+
+      if (!anchor) return
+
+      console.log('btn clicked')
+
+      await submitLoginForm(anchor)
+
+      window.location = anchor.getAttribute('href')
+    })
+
+    async function submitLoginForm(anchor) {
+      const loadingElement = document.getElementById(loginLoadingId)
+      const form = document.getElementById('login-form')
+
+      const formData = new FormData(form)
+      const email = formData.get('email')
+      const password = formData.get('password')
+
+      const credentials = {
+        email,
+        password
+      }
+
+      loading({ target: submitBtn, loadingElement, loading: true })
+
+      return new Promise((resolve, reject) => {
+        setTimeout(() => {
+          form.reset()
+
+          loading({ target: submitBtn, loadingElement, loading: false })
+          console.log(credentials)
+
+          resolve()
+        }, 2000)
+      })
+    }
+  }
 }
 
 export default new LoginPage({
   template,
   props: {
-    href: '/register',
-    value: 'Нет аккаунта?',
-    buttonId: loginBtnId,
-    registerBtnId,
-    loginBtnId,
     loadingId: loginLoadingId,
-    form
+    form,
+    LoginBtn: Button.template({
+      ...Button.props,
+      class: 'bg-primary white',
+      value: 'Войти',
+      href: '/',
+      id: loginBtnId
+    }),
+    RegisterBtn: Button.template({
+      ...Button.props,
+      class: 'white',
+      value: 'Регистрация',
+      href: '/register',
+      id: registerBtnId,
+      outline: true
+    })
   }
 })
